requirements.txt
py-doccle
python-dotenv
pathvalidate
requests